Optimize labels sample
