+++ app/components/AddLabelForm.vue
@@ -1,0 +1,66 @@
+<template>
+  <div
+    class="px-3 py-3 flex flex-row justify-between items-center">
+    <div class="mr-4 relative w-full">
+      <input
+        v-model="inputLabel"
+        type="text"
+        id="label"
+        class="block px-2.5 pb-2.5 pt-4 w-full min-w-max text-sm text-gray-900 bg-transparent rounded-lg border-1 border-gray-300 appearance-none dark:text-white dark:border-gray-600 dark:focus:border-blue-500 focus:outline-none focus:ring-0 focus:border-blue-600 peer"
+        :class="duplicate ? 'border-red-500 dark:border-red-500' : ''"
+        placeholder=""
+        list="definedLabels"
+        @keyup.enter="addLabel"/>
+        <datalist v-if="definedLabels" id="definedLabels">
+          <option v-for="label in definedLabels" :key="label" :value="label"></option>
+        </datalist>
+      <label
+        for="label"
+        class="absolute text-sm text-gray-500 dark:text-gray-400 duration-300 transform -translate-y-4 scale-75 top-2 z-10 origin-[0] bg-transparent peer-focus:bg-gray-100 peer-focus:dark:bg-gray-900 px-2 peer-focus:px-2 peer-focus:text-blue-600 peer-focus:dark:text-blue-500 peer-placeholder-shown:scale-100 peer-placeholder-shown:-translate-y-1/2 peer-placeholder-shown:top-1/2 peer-focus:top-2 peer-focus:scale-75 peer-focus:-translate-y-4 left-1">
+        Enter your label
+      </label>
+    </div>
+    <button
+      class="px-5 py-3 md:px-4 md:py-3 sm:px-2 sm:py-2 min-w-fit bg-blue-400 dark:bg-blue-700 hover:bg-blue-500 hover:dark:bg-blue-600 text-white dark:text-slate-200 rounded-md"
+      @click.prevent="addLabel"
+      :disabled="isLoadingState ?? false">
+      <span v-if="isLoadingState ?? false" class="flex flex-inline">
+        <svg aria-hidden="true" role="status" class="inline w-4 h-4 mt-1 mr-2 text-gray-200 animate-spin dark:text-gray-600" viewBox="0 0 100 101" fill="none" xmlns="http://www.w3.org/2000/svg">
+          <path d="M100 50.5908C100 78.2051 77.6142 100.591 50 100.591C22.3858 100.591 0 78.2051 0 50.5908C0 22.9766 22.3858 0.59082 50 0.59082C77.6142 0.59082 100 22.9766 100 50.5908ZM9.08144 50.5908C9.08144 73.1895 27.4013 91.5094 50 91.5094C72.5987 91.5094 90.9186 73.1895 90.9186 50.5908C90.9186 27.9921 72.5987 9.67226 50 9.67226C27.4013 9.67226 9.08144 27.9921 9.08144 50.5908Z" fill="currentColor"/>
+          <path d="M93.9676 39.0409C96.393 38.4038 97.8624 35.9116 97.0079 33.5539C95.2932 28.8227 92.871 24.3692 89.8167 20.348C85.8452 15.1192 80.8826 10.7238 75.2124 7.41289C69.5422 4.10194 63.2754 1.94025 56.7698 1.05124C51.7666 0.367541 46.6976 0.446843 41.7345 1.27873C39.2613 1.69328 37.813 4.19778 38.4501 6.62326C39.0873 9.04874 41.5694 10.4717 44.0505 10.1071C47.8511 9.54855 51.7191 9.52689 55.5402 10.0491C60.8642 10.7766 65.9928 12.5457 70.6331 15.2552C75.2735 17.9648 79.3347 21.5619 82.5849 25.841C84.9175 28.9121 86.7997 32.2913 88.1811 35.8758C89.083 38.2158 91.5421 39.6781 93.9676 39.0409Z" fill="#1C64F2"/>
+        </svg>
+        Processing...
+      </span>
+      <span v-else>Add label</span>
+    </button>
+  </div>
+</template>
+
+<script setup>
+  import { useState } from 'nuxt/app';
+  import { ref } from 'vue'
+  let inputLabel = ref('')
+  const useLabels = () => useState('labels', () => { return [] })
+  const duplicate = ref(false)
+
+  // Official defined list of defined labels
+  const useDefinedLabels = () => useState('definedLabels', () => { return [] })
+  const definedLabels = useDefinedLabels()
+
+
+  const addLabel =() =>{
+    const labels = useLabels()
+    const label = inputLabel.value.trim()
+    if (label && label.length > 0) {
+      if (labels.value.includes(label)) {
+        if (isDev()) console.info('label duplicated!: ' + label)
+        duplicate.value = true
+      } else {
+        labels.value.push(label)
+        duplicate.value = false
+        inputLabel.value = ''
+      }
+    }
+  }
+
+</script>
